refactor(mainwindow): split responsibilities into focused modules
Extract connection/session/TOFU and quick-site persistence flow into MainWindowConnection.cpp.

Move local filesystem operations to MainWindowLocalOps.cpp and remote SFTP operations to MainWindowRemoteOps.cpp.

Move transfer queue, remote prescan, and drag-and-drop handling to MainWindowTransfers.cpp, leaving MainWindow.cpp as UI orchestrator.

Update CMake target sources to compile the new MainWindow translation units.
